Add files via upload
This game creates a 5 digit random number. It then allows the player 5 attempts to guess that number. A GREEN dot is displayed and means that  the number is right & in the right place.\A RED dot means the number is right but in the wrong place. A BLACK dot means the number is wrong.